Updated variable indices in other implementations. Need to modify tests.
