+++ smol/solver/upper_bound/constraints.py
@@ -6,7 +6,7 @@ import cvxpy as cp
 import numpy as np
 from cvxpy import Expression
 from cvxpy.constraints.constraint import Constraint
-from numpy.typing import ArrayLike
+from pymatgen.core import Structure
 
 from smol.moca.composition import CompositionSpace
 from smol.moca.ensemble import Sublattice
@@ -31,21 +31,23 @@ def get_upper_bound_normalization_constraints(
     Returns:
         List[Constraint].
     """
-    return [
-        cp.sum(variables[indices]) == 1
-        for indices in variable_indices
-        if len(indices) > 0
-    ]
+    constraints = []
+    for site_indices in variable_indices:
+        site_variable_indices = np.array(site_indices, dtype=int)
+        site_variable_indices = site_variable_indices[site_variable_indices >= 0]
+        if len(site_variable_indices) > 0:
+            constraints.append(cp.sum(variables[site_variable_indices]) == 1)
+    return constraints
 
 
 def get_upper_bound_composition_space_constraints(
     sublattices: List[Sublattice],
     variables: cp.Variable,
     variable_indices: List[List[int]],
+    processor_structure: Structure,
     other_constraints: List[
         Union[str, Tuple[Union[dict, List[dict], List[Number]], Number, str]]
     ] = None,
-    initial_occupancy: ArrayLike = None,
 ) -> List[Constraint]:
     """Get constraints on species composition in CompositionSpace.
 
@@ -58,22 +60,24 @@ def get_upper_bound_composition_space_constraints(
             cvxpy variables storing the ground-state result.
         variable_indices(list[list[int]]):
             List of variable indices corresponding to each active site index and
-            species indices in its site space.
+            index of species in its site space. Inactive sites will be marked by
+            either -1 or -2. See documentation in solver.upper_bound.variables.
+        processor_structure(Structure):
+            The supercell structure stored in a processor's structure attribute.
+            The sub-lattices must match the processor structure, or they must be the result
+            of splitting with the initial_occupancy. See smol.moca.sublattice for the
+            explanation of splitting a sub-lattice.
         other_constraints(list): optional
             Other constraints to set in composition space. See moca.composition
             for detailed description.
             Note that constraints are now to be satisfied with the number of sites
             in sub-lattices of the supercell instead of the primitive cell.
-        initial_occupancy(ArrayLike): optional
-            An initial occupancy used to set the occupancy of manually restricted
-            sites that may have more than one allowed species.
-            Must be provided if any site has been manually restricted.
     Returns:
         list[Constraint]: Constraints corresponding to the given composition space.
     """
     # Get the variable indices corresponding to each dimension in "counts" format.
     variables_per_component = get_variable_indices_for_each_composition_component(
-        sublattices, variable_indices, initial_occupancy
+        sublattices, variable_indices, processor_structure
     )
 
     # Create the composition space.
@@ -178,8 +182,8 @@ def get_upper_bound_fixed_composition_constraints(
     sublattices: List[Sublattice],
     variables: cp.Variable,
     variable_indices: List[List[int]],
+    processor_structure: Structure,
     fixed_composition: List[int],
-    initial_occupancy: ArrayLike = None,
 ) -> List[Constraint]:
     """Fix the count of species in the super-cell.
 
@@ -191,22 +195,24 @@ def get_upper_bound_fixed_composition_constraints(
             cvxpy variables storing the ground-state result.
         variable_indices(list[list[int]]):
             List of variable indices corresponding to each active site index and
-            species indices in its site space.
+            index of species in its site space. Inactive sites will be marked by
+            either -1 or -2. See documentation in solver.upper_bound.variables.
+        processor_structure(Structure):
+            The supercell structure stored in a processor's structure attribute.
+            The sub-lattices must match the processor structure, or they must be the result
+            of splitting with the initial_occupancy. See smol.moca.sublattice for the
+            explanation of splitting a sub-lattice.
         fixed_composition(list[int]):
             Amount of each species to be fixed in the SUPER-cell, in CompositionSpace
             "counts" format. You are fully responsible for setting the order of species
             in the list correctly as would be generated by a CompositionSpace.
-        initial_occupancy(ArrayLike): optional
-            An initial occupancy used to set the occupancy of manually restricted
-            sites that may have more than one allowed species.
-            Must be provided if any site has been manually restricted.
     Return:
         list[Constraint]: Constraints corresponding to a fixed composition on each
         sub-lattice.
     """
     # Get the variable indices corresponding to each dimension in "counts" format.
     variables_per_component = get_variable_indices_for_each_composition_component(
-        sublattices, variable_indices, initial_occupancy
+        sublattices, variable_indices, processor_structure
     )
 
     constraints = []
